// packages/renderer-pixi/src/render/generatedAssetJson.ts
export const pickupAnims = {
  "sheets": {
    "small": "sheal.png",
    "large": "heal.png",
    "sammo": "sammo.png",
    "ammo": "ammo.png"
  },
  "actors": {
    "small": {
      "sheet": "sheal.png",
      "animations": {
        "falling": {
          "loop": false,
          "speed": 1,
          "frames": [
            {
              "region": [0, 0, 16, 16],
              "duration": 1
            }
          ]
        },
        "idle": {
          "loop": true,
          "speed": 24,
          "frames": [
            {
              "region": [16, 0, 16, 16],
              "duration": 1
            },
            {
              "region": [0, 16, 16, 16],
              "duration": 1
            },
            {
              "region": [16, 16, 16, 16],
              "duration": 1
            }
          ]
        }
      }
    },
    "large": {
      "sheet": "heal.png",
      "animations": {
        "idle": {
          "loop": true,
          "speed": 24,
          "frames": [
            {
              "region": [16, 0, 16, 16],
              "duration": 1
            },
            {
              "region": [0, 16, 16, 16],
              "duration": 1
            },
            {
              "region": [16, 16, 16, 16],
              "duration": 1
            }
          ]
        },
        "falling": {
          "loop": false,
          "speed": 1,
          "frames": [
            {
              "region": [0, 0, 16, 16],
              "duration": 1
            }
          ]
        }
      }
    },
    "sammo": {
      "sheet": "sammo.png",
      "animations": {
        "falling": {
          "loop": false,
          "speed": 1,
          "frames": [
            {
              "region": [0, 0, 16, 16],
              "duration": 1
            }
          ]
        },
        "idle": {
          "loop": true,
          "speed": 24,
          "frames": [
            {
              "region": [16, 0, 16, 16],
              "duration": 1
            },
            {
              "region": [0, 16, 16, 16],
              "duration": 1
            },
            {
              "region": [16, 16, 16, 16],
              "duration": 1
            }
          ]
        }
      }
    },
    "ammo": {
      "sheet": "ammo.png",
      "animations": {
        "falling": {
          "loop": false,
          "speed": 1,
          "frames": [
            {
              "region": [0, 0, 16, 16],
              "duration": 1
            }
          ]
        },
        "idle": {
          "loop": true,
          "speed": 24,
          "frames": [
            {
              "region": [16, 0, 16, 16],
              "duration": 1
            },
            {
              "region": [0, 16, 16, 16],
              "duration": 1
            },
            {
              "region": [16, 16, 16, 16],
              "duration": 1
            }
          ]
        }
      }
    }
  }
}